feat: 예약대기 승인 메서드 ReserveTicket 도메인으로 이동
