+++ src/main/java/roomescape/service/waiting/WaitingService.java
@@ -1,18 +1,15 @@
 package roomescape.service.waiting;
 
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.member.Member;
 import roomescape.domain.reservationtime.ReservationTime;
 import roomescape.domain.theme.Theme;
 import roomescape.domain.waiting.Waiting;
 import roomescape.domain.waiting.WaitingWithRank;
 import roomescape.dto.reservation.AddReservationDto;
-import roomescape.dto.waiting.ApplyWaitingRequestDto;
 import roomescape.repository.waiting.WaitingRepository;
 import roomescape.service.member.MemberService;
 import roomescape.service.reservationtime.ReservationTimeService;
-import roomescape.service.reserveticket.ReserveTicketService;
 import roomescape.service.theme.ThemeService;
 
 import java.time.LocalDate;
@@ -27,14 +24,12 @@ public class WaitingService {
     private final MemberService memberService;
     private final ReservationTimeService reservationTimeService;
     private final ThemeService themeService;
-    private final ReserveTicketService reserveTicketService;
 
-    public WaitingService(WaitingRepository waitingRepository, MemberService memberService, ReservationTimeService reservationTimeService, ThemeService themeService, ReserveTicketService reserveTicketService) {
+    public WaitingService(WaitingRepository waitingRepository, MemberService memberService, ReservationTimeService reservationTimeService, ThemeService themeService) {
         this.waitingRepository = waitingRepository;
         this.memberService = memberService;
         this.reservationTimeService = reservationTimeService;
         this.themeService = themeService;
-        this.reserveTicketService = reserveTicketService;
     }
 
     public long addWaiting(AddReservationDto newReservationDto, Long memberId) {
@@ -83,14 +78,5 @@ public class WaitingService {
 
     public void deleteWaiting(Long id) {
         waitingRepository.deleteById(id);
-    }
-
-    @Transactional
-    public Long apply(ApplyWaitingRequestDto applyWaitingRequestDto) {
-        Long waitingId = applyWaitingRequestDto.id();
-        Waiting waiting = getWaitingById(waitingId);
-        Long reservationId = reserveTicketService.addReservation(new AddReservationDto(waiting.getDate(), waiting.getTime().getId(), waiting.getTheme().getId()), waiting.getMember().getId());
-        deleteWaiting(waitingId);
-        return reservationId;
     }
 }
